Fixed map y-axis, included distance calibrated data
Modified the sorting program so that it makes a .txt file of the stars to which distance is known, which I assume will help with calibration of some sort further down the line.
I also changed the sorting program so that it's more strict with low parallax angles, which will give better results (changed threshold from >0.001 to >0.002).