X-Ray Analysis, Fuzzy Clusters, reformatting
Includes a lot of new code for X-Ray analysis (and includes a rudimentary distance to velocity graph which roughly shows Hubble's constant with an excellent fit!). Also reformatted lots of code, and included a new folder which contains a .txt file for each fuzzy cluster complete with the properties of each galaxy in the cluster. I plan on working on all of this more throughout the week, and am just committing it now so that I can work on it on my other computer :)